feat: 기존 프로그램 정상 종료시 FileMemberRepository 스레드 종료 로직 -> 강제종료시에도 실행되도록 shutdownHook 추가

+++ src/main/java/org/sopt/service/MemberService.java
@@ -46,6 +46,7 @@ public class MemberService {
     }
 
     public void deleteMember(Long id) {
+        if(!memoryMemberRepository.isExistId(id)) throw new IllegalArgumentException("존재하지 않은 회원 ID입니다.");
         memoryMemberRepository.deleteById(id);
     }
 }
